requirements.txt
requests
beautifulsoup4
google
dnspython
yagooglesearch